refactor(grid): 优化目录结构
